refactor: migrate to bun

+++ src/commands/restart.ts
@@ -4,6 +4,9 @@ const RestartCommand: Command = {
   alias: "restart",
   permission: PermissionsBitField.Flags.Administrator,
   exec: async (msg, args) => {
+    if (!msg.channel.isSendable()) {
+      return;
+    }
     msg.channel.send("Gonna kill myself, I hope someone will catch my fall...");
     setTimeout(() => {
       process.exit();
